generationPipeline/requirements.txt
sentence-transformers
pika #rabbitmq
google-genai
pinecone
dotenv

